share one domain-set walk between the citation snapshot and the categorizer
The snapshot route carried its own copy of "equals, or is a subdomain
of, any entry in this set". inDomainSet moves to the client-safe module
so both edges answer it the same way.
